feat(seo): wave 1 — indexation correctness
The codebase audit flagged indexation as the top-of-funnel problem: a
handful of duplicate-content URLs were indexable, the title template
double-stacked the brand suffix on blog and compare detail pages, the
blog post body was being client-rendered into an empty initial HTML
shell (soft-404 risk to crawlers), and `/compare/[competitor]` emitted
two competing `SoftwareApplication` schemas on the same URL.

This first wave addresses the indexation gaps without touching brand
or rich-result expansion (those land in subsequent waves).

Changes:

- Stripe-result and Stripe-checkout completion pages get an explicit
  `noindex,nofollow` (`/stripe/cancel`, `/stripe/success`,
  `/pricing/complete`). They duplicate `/pricing/cancel|success` and
  have no SEO value. `robots.ts` also disallows `/stripe` and
  `/pricing/complete` at the protocol level for defense in depth.
- `/pricing/complete/page.tsx` was `'use client'`, so the metadata
  export had no effect. Split into a server `page.tsx` that exports
  metadata + renders the client `complete-client.tsx`.
- Blog detail metadata used `title: 'X | TenantFlow Blog'` (a string),
  which the parent layout's `'%s | TenantFlow'` template then suffixed,
  producing `"X | TenantFlow Blog | TenantFlow"`. Switched to
  `title.absolute` to opt out of the template. Same fix on
  `/compare/[competitor]/page.tsx`.
- Blog detail OG had hard-coded `width: 1080, height: 607` — below
  the 1200x630 floor that LinkedIn / Slack / Discord scrapers require
  and not a true fact about external `featured_image` URLs. Dropped
  the dimension hints and let the platform sniff. Added
  `modifiedTime`, `twitter:site`, `twitter:creator` (the audit flagged
  these as missing).
- Compare detail used `og:type: website` for what is effectively
  comparison-article content. Switched to `article` and added
  `siteName` (was inheriting from parent — fragile).
- The biggest indexation fix: the blog post body was rendered by a
  `'use client'` component that called `useBlogBySlug` after hydration.
  Initial HTML had only a skeleton + JSON-LD scripts. Refactored
  `BlogPostPage` to take `post` + `slug` as props and dropped the
  `useBlogBySlug` hook. The server `page.tsx` already had the post
  via `getBlogPost()` (cached + shared with `generateMetadata`); now
  it passes the post through. Initial HTML now contains the h1,
  excerpt, hero image, byline, and category — the fields crawlers
  parse for Article rich-result eligibility.
- `Article` schema dropped its `new Date().toISOString()` fallback
  for `datePublished`. Faking freshness is the kind of thing that
  teaches Google to discount a site's lastmod signal site-wide. The
  schema only emits when the post has a real `published_at`.
- Compare detail's two `SoftwareApplication` schemas (one for
  TenantFlow, one for the competitor) split rich-result eligibility
  on the same URL. The competitor isn't our product. Dropped both —
  the root layout already emits a sitewide `SoftwareApplication` for
  TenantFlow, and the comparison itself is captured by `BreadcrumbList`
  + on-page content.
- `robots.ts` now declares per-bot rules for the canonical AI
  user-agents (`GPTBot`, `OAI-SearchBot`, `ChatGPT-User`,
  `Google-Extended`, `ClaudeBot`, `Claude-User`, `Claude-SearchBot`,
  `Applebot-Extended`, `CCBot`, `PerplexityBot`, `Perplexity-User`,
  `Amazonbot`). Marketing surface stays allowed by default — the
  same private paths as the wildcard rule are denied per bot — so
  flipping a single line opts a specific crawler out without
  rewriting the wildcard.
- Compare detail also gains "Compare TenantFlow to other tools"
  cross-links to sibling competitor pages. Each comparison page used
  to be an isolated leaf with no internal link to the other two.

Tests touched: blog detail tests passed `post` via mocks of
`useBlogBySlug`. Updated to render with the new prop shape directly,
and removed the loading-skeleton + not-found tests because the server
now resolves the post before the client component renders (and calls
`notFound()` for missing slugs in `page.tsx`). Unit suite still
passes — 99,866 tests.

[hudsor01]
