+++ src/app/stripe/success/page.tsx
@@ -1,8 +1,15 @@
+import type { Metadata } from 'next'
 import { Button } from '#components/ui/button'
 import { PageLayout } from '#components/layout/page-layout'
 import { CheckCircle } from 'lucide-react'
 import Link from 'next/link'
 import { Suspense } from 'react'
+
+// Transactional Stripe-result page — no SEO value, duplicates `/pricing/success`.
+// robots.ts also disallows `/stripe/*` for defense in depth.
+export const metadata: Metadata = {
+	robots: { index: false, follow: false },
+}
 
 function SuccessContent() {
 	return (
